+++ cmd/launcher/launcher.go
@@ -45,7 +45,7 @@ import (
 	osqueryInstanceHistory "github.com/kolide/launcher/pkg/osquery/runtime/history"
 	"github.com/kolide/launcher/pkg/service"
 	"github.com/kolide/launcher/pkg/traces/exporter"
-	"github.com/kolide/launcher/pkg/windows/eventwatcher"
+	"github.com/kolide/launcher/pkg/windows/powereventwatcher"
 	"github.com/oklog/run"
 
 	"go.etcd.io/bbolt"
@@ -210,7 +210,7 @@ func runLauncher(ctx context.Context, cancel func(), opts *launcher.Options) err
 		close(sigChannel)
 	})
 
-	powerEventWatcher, err := eventwatcher.New(log.With(logger, "component", "power_event_watcher"))
+	powerEventWatcher, err := powereventwatcher.New(log.With(logger, "component", "power_event_watcher"))
 	if err != nil {
 		level.Debug(logger).Log("msg", "could not init power event watcher", "err", err)
 	} else {
