+++ pkg/windows/powereventwatcher/power_event_watcher_windows.go
@@ -1,7 +1,7 @@
 //go:build windows
 // +build windows
 
-package eventwatcher
+package powereventwatcher
 
 import (
 	"encoding/xml"
